+++ src/main/java/com/hexvane/orbisorigins/species/SpeciesJsonCodec.java
@@ -1,7 +1,7 @@
 package com.hexvane.orbisorigins.species;
 
 import com.hypixel.hytale.math.shape.Box;
-import com.hypixel.hytale.math.vector.Vector3d;
+import org.joml.Vector3d;
 import com.nimbusds.jose.shaded.gson.Gson;
 import com.nimbusds.jose.shaded.gson.GsonBuilder;
 import com.nimbusds.jose.shaded.gson.JsonDeserializationContext;